Enable sharing the cache between requests
The code is now much less generic. I intend to make it more generic for
any Cache and any Content. Hopefully, it's possible/not too difficult.

One upside, is that changing the type of the Cache, is just a matter of
changing the type aliasing of SharedCache. For example, instead of using
MemoryCache, it could be using RemoteCache (currently, not existing,
tho).